Make Quotable LambdaOp Invocation Handler an inner anonymous class
